+++ lib/tello/controller/command_builder.ex
@@ -39,7 +39,7 @@ defmodule Tello.Controller.CommandBuilder do
     "ccw #{degree}"
   end
 
-  def control(:go, {x, y, z}, speed, mission_pad_id)
+  def control(:go, {x, y, z}, speed, mission_pad_id \\ nil)
       when x in -500..500 and y in -500..500 and z in -500..500 and
              speed in 10..100 do
     "go #{x} #{y} #{z} #{speed} #{mission_pad_id}"