+++ lib/tello/controller/controller.ex
@@ -1,6 +1,8 @@
 defmodule Tello.Controller do
   @moduledoc """
   GenServer to connect to a Tello.
+
+  Uses `Tello.start/1` to start a client.
   """
 
   use GenServer
@@ -8,6 +10,7 @@ defmodule Tello.Controller do
 
   alias Tello.Controller.{State, CommandBuilder}
 
+  @type t :: module()
   @type init_arg :: [
           ip: :inet.ip_address(),
           port: :inet.port_number(),
@@ -16,12 +19,14 @@ defmodule Tello.Controller do
 
   # Server (callbacks)
 
+  @doc false
   @spec start_link(uid: reference(), arg: init_arg()) :: :ignore | {:error, any} | {:ok, pid}
   def start_link(uid: uid, arg: arg) do
     GenServer.start_link(__MODULE__, arg, name: :"#{__MODULE__}.#{uid}")
   end
 
   @impl true
+  @doc false
   def init(ip: ip, port: port) do
     init(ip: ip, port: port, receiver: nil)
   end
@@ -68,7 +73,13 @@ defmodule Tello.Controller do
         {:udp, _socket, _ip, _port, data},
         %State{receiver_module: receiver_module} = state
       ) do
-    receiver_module.receive_message(data)
+    if function_exported?(receiver_module, :receive_message, 1) do
+      receiver_module.receive_message(data)
+    else
+      Logger.warn(
+        "Please implement `receive_message/1` for the custom `Tello.Controller.Receiver` handler."
+      )
+    end
 
     {:noreply, state}
   end
@@ -79,6 +90,7 @@ defmodule Tello.Controller do
   @doc """
   Enable Tello's SDK mode.
   """
+  @spec enable(pid) :: :ok | {:error, any()}
   def enable(controller) do
     GenServer.call(controller, {:send, "command"})
   end
@@ -86,6 +98,7 @@ defmodule Tello.Controller do
   @doc """
   Auto takeoff.
   """
+  @spec takeoff(pid) :: :ok | {:error, any()}
   def takeoff(controller) do
     GenServer.call(controller, {:send, "takeoff"})
   end
@@ -93,6 +106,7 @@ defmodule Tello.Controller do
   @doc """
   Auto landing.
   """
+  @spec land(pid) :: :ok | {:error, any()}
   def land(controller) do
     GenServer.call(controller, {:send, "land"})
   end
@@ -100,6 +114,7 @@ defmodule Tello.Controller do
   @doc """
   Stop motors immediately.
   """
+  @spec emergency(pid) :: :ok | {:error, any()}
   def emergency(controller) do
     GenServer.call(controller, {:send, "emergency"})
   end
@@ -107,6 +122,7 @@ defmodule Tello.Controller do
   @doc """
   Hover in the air.
   """
+  @spec stop(pid) :: :ok | {:error, any()}
   def stop(controller) do
     GenServer.call(controller, {:send, "stop"})
   end
@@ -126,6 +142,7 @@ defmodule Tello.Controller do
     @doc """
     Fly #{command} for `distance` cm.
     """
+    @spec unquote(command)(pid, integer) :: :ok | {:error, any()}
     def unquote(command)(controller, distance) do
       command = CommandBuilder.control(unquote(command), distance)
 
@@ -142,6 +159,7 @@ defmodule Tello.Controller do
   - `forward`
   - `back`
   """
+  @spec flip(pid, :back | :forward | :left | :right) :: :ok | {:error, any()}
   def flip(controller, direction)
       when direction in [:left, :right, :forward, :back] do
     command = CommandBuilder.control(:flip, direction)
@@ -165,8 +183,8 @@ defmodule Tello.Controller do
 
   If `mission_pad_id` is set, the coordinates will be set of the Mission Pad.
   """
-  @spec go(pid(), coordinate, integer(), mission_pad_id) :: :ok | {:error, any()}
-  def go(controller, _coordinate = {x, y, z}, speed, mission_pad_id \\ nil)
+  @spec go(pid(), coordinate, integer(), mission_pad_id() | nil) :: :ok | {:error, any()}
+  def go(controller, {x, y, z}, speed, mission_pad_id \\ nil)
       when x in -500..500 and y in -500..500 and z in -500..500 and
              speed in 10..100 and
              mission_pad_id in [:m1, :m2, :m3, :m4, :m5, :m6, :m7, :m8] do
@@ -180,9 +198,17 @@ defmodule Tello.Controller do
   and recognize coordinates `0,0,z` of Mission Pad `second_mission_pad_id`
   and rotate to the `yaw` value after hovering at the coordinates.
   """
+  @spec jump(
+          pid,
+          coordinate(),
+          integer(),
+          integer(),
+          mission_pad_id(),
+          mission_pad_id()
+        ) :: :ok | {:error, any()}
   def jump(
         controller,
-        _coordinate = {x, y, z},
+        {x, y, z},
         speed,
         yaw,
         first_mission_pad_id,
@@ -209,6 +235,8 @@ defmodule Tello.Controller do
   Fly at a curve according to the two given coordinates of the Mission Pad `mission_pad_id`
   at `speed` (cm/s).
   """
+  @spec curve(pid, coordinate(), coordinate(), integer(), mission_pad_id() | nil) ::
+          :ok | {:error, any()}
   def curve(
         controller,
         _coordinate1 = {x1, y1, z1},
@@ -228,6 +256,7 @@ defmodule Tello.Controller do
   @doc """
   Set speed to `speed` cm/s
   """
+  @spec set_speed(pid(), integer()) :: :ok | {:error, any()}
   def set_speed(controller, speed) when speed in 10..100 do
     command = CommandBuilder.set(:speed, speed)
 
@@ -237,6 +266,13 @@ defmodule Tello.Controller do
   @doc """
   Set remote controller via four channels
   """
+  @spec set_remote_controller(
+          pid(),
+          {:left, integer()} | {:right, integer()},
+          {:backward, integer()} | {:forward, integer()},
+          {:down, integer()} | {:up, integer()},
+          integer()
+        ) :: :ok | {:error, any()}
   def set_remote_controller(
         controller,
         channel_left_right = {a_direction, a_value},
@@ -260,6 +296,7 @@ defmodule Tello.Controller do
   @doc """
   Set Wi-Fi SSID and password
   """
+  @spec set_wifi(pid(), String.t(), String.t()) :: :ok | {:error, any()}
   def set_wifi(controller, ssid, password) do
     command = CommandBuilder.set(:wifi, ssid, password)
 
@@ -269,6 +306,7 @@ defmodule Tello.Controller do
   @doc """
   Set Mission Pad detection to on/off
   """
+  @spec set_mission_pad_detection(pid(), :off | :on) :: :ok | {:error, any()}
   def set_mission_pad_detection(controller, toggle)
       when toggle in [:on, :off] do
     command = CommandBuilder.set(:mission_pad_detection, toggle)
@@ -279,6 +317,10 @@ defmodule Tello.Controller do
   @doc """
   Set Mission Pad detection mode
   """
+  @spec set_mission_pad_detection_mode(
+          pid(),
+          :both | :downward | :forward
+        ) :: :ok | {:error, any()}
   def set_mission_pad_detection_mode(controller, mode)
       when mode in [:downward, :forward, :both] do
     command = CommandBuilder.set(:mission_pad_detection_mode, mode)
@@ -289,6 +331,7 @@ defmodule Tello.Controller do
   @doc """
   Set the Tello to station mode, and connect to the access point.
   """
+  @spec connect_to_ap(pid(), String.t(), String.t()) :: :ok | {:error, any()}
   def connect_to_ap(controller, ssid, password) do
     command = CommandBuilder.set(:ap, ssid, password)
 
@@ -298,6 +341,7 @@ defmodule Tello.Controller do
   @doc """
   Get current speed (cm/s).
   """
+  @spec get_speed(pid) :: :ok | {:error, any()}
   def get_speed(controller) do
     command = CommandBuilder.read(:speed)
 
@@ -307,6 +351,7 @@ defmodule Tello.Controller do
   @doc """
   Get current battery percentage.
   """
+  @spec get_battery(pid) :: :ok | {:error, any()}
   def get_battery(controller) do
     command = CommandBuilder.read(:battery)
 
@@ -316,6 +361,7 @@ defmodule Tello.Controller do
   @doc """
   Get current flight time.
   """
+  @spec get_time(pid) :: :ok | {:error, any()}
   def get_time(controller) do
     command = CommandBuilder.read(:time)
 
@@ -325,6 +371,7 @@ defmodule Tello.Controller do
   @doc """
   Get Wi-Fi SNR.
   """
+  @spec get_wifi_snr(pid) :: :ok | {:error, any()}
   def get_wifi_snr(controller) do
     command = CommandBuilder.read(:wifi_snr)
 
@@ -334,6 +381,7 @@ defmodule Tello.Controller do
   @doc """
   Get Tello SDK version
   """
+  @spec get_sdk_version(pid) :: :ok | {:error, any()}
   def get_sdk_version(controller) do
     command = CommandBuilder.read(:sdk_version)
 
@@ -343,6 +391,7 @@ defmodule Tello.Controller do
   @doc """
   Get Tello serial number
   """
+  @spec get_serial_number(pid) :: :ok | {:error, any()}
   def get_serial_number(controller) do
     command = CommandBuilder.read(:serial_number)