+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## main
 
-
+- Rename `Tello.Command` to `Tello.Controller`.
 - Added `Tello.StatusListener` to receive status update from Tello.
 - Allow to attach custom receiver to `Tello.Controller` to receive message form Tello, see `Tello.Controller.Receiver` for more detail.